FIX: dinero exacto e historial de tarifas docentes
Los importes vivían en `double precision`: no se guardaban exactos y las sumas
arrastraban el error, que en el saldo de un alumno se lee como un descuadre sin
causa. Pasan a NUMERIC(12,2) y a Decimal de extremo a extremo (pagos, facturas,
cuotas, tarifas). Hacia el frontend siguen viajando como número JSON, así que no
cambia nada de su lado; en la auditoría se guardan como cadena exacta.

La nómina liquidaba cualquier fecha con la tarifa *actual* del profesor, de modo
que subírsela reescribía hacia atrás meses ya cerrados y pagados por otro
importe. `teacher_rates` guarda desde cuándo rige cada tarifa y cada clase se
paga con la que estaba vigente ese día; la migración conserva la tarifa actual
como primera fila histórica para no dejar en cero lo ya liquidado.

+++ backend/app/models/invoice.py
@@ -2,7 +2,9 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, Float, ForeignKey, String, func
+from decimal import Decimal
+
+from sqlalchemy import DateTime, ForeignKey, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -20,7 +22,7 @@ class Invoice(Base):
         ForeignKey("enrollments.id", ondelete="CASCADE"), index=True
     )
     code: Mapped[str] = mapped_column(String(32), unique=True, index=True)
-    total_amount: Mapped[float] = mapped_column(Float)
+    total_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2))
     issued_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now()
     )
